Set Provider's Key-Value Form return messages' Content-Type header to not 'text/html'.
Free hosting services (such as those offered by GoDaddy) add a trailing <script> tag to responses with a Content-Type of text/html.  But since this tag invalidates the format of Key-Value Form, and since KVF isn't html anyway, we set content-type to something else, which allows a DotNetOpenId Provider to be hosted on a free hosting site by GoDaddy without it corrupting the response messages.
New Content-Type for KVF is "application/x-openid-kvf", because text/plain was insufficient to avoid the trailing <script> tag.