Fixed Token to drop a hint about its signature so it can always deserialize properly.
Before, token deserialization would assume that if app state is available then it was available during serialization as well and would expect a signature.
Now token serialization uses an extra byte in the front to indicate whether there is a signature present so that it can do the right thing in any case.