OpenIdAjaxTextBox rewritten to derive from OpenIdRelyingPartyAjaxControlBase.
A great deal of refactoring went into this, but of C# and Javascript code.  The upside to this is now we can write many more AJAX controls (auto-login, for example) with minimal code.

Note that the OpenIdAjaxTextBox control is not yet back up to par with the prior one.  Postbacks don't yet properly deserialize the 'authenticated' state of the page.