Resource owner password grant method rename
Renamed IAuthorizationServerHost.IsResourceOwnerCredentialValid to
TryAuthorizeResourceOwnerCredentialGrant so that authorization servers
are prepared to approve refresh tokens that may be issued as a result of
a resource owner password grant.

This also removes some of the "validation" that wasn't really doing
anything useful for resource owner password grant types.

Toward an eventual fix for #138