Fixes a bug in posting to HTML forms in Firefox with the InfoCard selector.
If the idselector control on a page with an other control that raises a postback, then the selector is opening when using firefox with "openinfocard identity selector" addon.
To fix this problem, you have to move the object tag with the identity object outside the form tag in the page.

Thanks to Philipp Jenni <philipp.jenni@gmx.ch> who contributed this fix.